# High-Performance Particle System Rewrite