… source/MonoGame.Extended/Particles/Data/ParticleReleaseParameters.cs
… source/MonoGame.Extended/Particles/Data/ParticleVector2Parameter.cs
… source/MonoGame.Extended/Particles/FastRandomExtensions.cs
… source/MonoGame.Extended/Particles/LineSegment.cs
… source/MonoGame.Extended/Particles/Modifiers/AgeModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/Containers/CircleContainerModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/Containers/RectangleContainerModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/Containers/RectangleLoopContainerModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/DragModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/ColorInterpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/HueInterpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/Interpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/InterpolatorOfT.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/OpacityInterpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/ScaleInterpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/VelocityInterpolator
… source/MonoGame.Extended/Particles/Modifiers/Interpolators/VelocityInterpolator.cs
… source/MonoGame.Extended/Particles/Modifiers/LinearGravityModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/Modifier.cs
… source/MonoGame.Extended/Particles/Modifiers/ModifierExecutionStrategy.cs
… source/MonoGame.Extended/Particles/Modifiers/OpacityFastFadeModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/RotationModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/VelocityColorModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/VelocityModifier.cs
… source/MonoGame.Extended/Particles/Modifiers/VortexModifier.cs
… source/MonoGame.Extended/Particles/Particle.cs
… source/MonoGame.Extended/Particles/ParticleEffect.cs
… source/MonoGame.Extended/Particles/ParticleEffectReader.cs
… source/MonoGame.Extended/Particles/ParticleEffectWriter.cs
… source/MonoGame.Extended/Particles/ParticleEmitter.cs