This PR introduces a completely rewritten particle system for MonoGame.Extended that delivers significant performance improvements and enhanced functionality.

## Performance Improvements

The new implementation shows substantial performance gains, particularly for heavy workloads:

| Scenario | Particle Count | Current Implementation | New Implementation | **Improvement**  |
| -------- | -------------- | ---------------------- | ------------------ | ---------------- |
| Heavy    | 100            | 629.1 μs               | 353.4 μs           | **43.8% faster** |
| Heavy    | 500            | 627.9 μs               | 351.2 μs           | **44.1% faster** |
| Heavy    | 1000           | 633.3 μs               | 361.7 μs           | **42.9% faster** |
| Heavy    | 2500           | 637.8 μs               | 371.5 μs           | **41.8% faster** |
| Heavy    | 5000           | 638.1 μs               | 353.5 μs           | **44.6% faster** |

Light and medium workloads maintain similar performance while providing better consistency and lower memory allocations. Full benchmarks found below

> [!NOTE]
> Benchmarks were done using:
>
> - A "light" effect which contains a single modifier
> - A "medium" effect which contains two modifiers
> - A "heavy" effect which contains five modifiers
>
> Each effect is run for 60 frames, triggering particles every 10 frames at a simulated 60fps

### Current Implementation Benchmark Results

| Method                               | ParticleCount |     Mean |    Error |   StdDev |   Median | Allocated |
| ------------------------------------ | ------------- | -------: | -------: | -------: | -------: | --------: |
| SustainedPerformance_Light_60Frames  | 100           | 126.8 us |  2.28 us |  5.10 us | 125.2 us |     112 B |
| SustainedPerformance_Medium_60Frames | 100           | 176.5 us |  2.96 us |  6.37 us | 175.1 us |     400 B |
| SustainedPerformance_Heavy_60Frames  | 100           | 629.1 us | 12.52 us | 16.28 us | 628.5 us |      64 B |
| SustainedPerformance_Light_60Frames  | 500           | 128.7 us |  4.18 us | 11.23 us | 124.5 us |     400 B |
| SustainedPerformance_Medium_60Frames | 500           | 177.4 us |  3.53 us |  7.67 us | 174.1 us |     112 B |
| SustainedPerformance_Heavy_60Frames  | 500           | 627.9 us |  9.73 us | 10.00 us | 626.0 us |      64 B |
| SustainedPerformance_Light_60Frames  | 1000          | 127.8 us |  3.45 us |  9.66 us | 124.1 us |      64 B |
| SustainedPerformance_Medium_60Frames | 1000          | 173.7 us |  2.96 us |  7.90 us | 171.0 us |      64 B |
| SustainedPerformance_Heavy_60Frames  | 1000          | 633.3 us | 12.40 us | 15.68 us | 632.9 us |     400 B |
| SustainedPerformance_Light_60Frames  | 2500          | 143.4 us |  6.82 us | 20.01 us | 134.1 us |     112 B |
| SustainedPerformance_Medium_60Frames | 2500          | 184.4 us |  3.65 us |  6.94 us | 182.9 us |     112 B |
| SustainedPerformance_Heavy_60Frames  | 2500          | 637.8 us | 11.86 us | 28.18 us | 631.7 us |     112 B |
| SustainedPerformance_Light_60Frames  | 5000          | 143.4 us |  6.43 us | 18.77 us | 138.8 us |     400 B |
| SustainedPerformance_Medium_60Frames | 5000          | 179.7 us |  3.59 us |  8.81 us | 177.6 us |      64 B |
| SustainedPerformance_Heavy_60Frames  | 5000          | 638.1 us | 12.76 us | 21.67 us | 630.4 us |     112 B |

### New Implementation Benchmark Results

| Method                               | ParticleCount |     Mean |    Error |   StdDev |   Median | Allocated |
| ------------------------------------ | ------------- | -------: | -------: | -------: | -------: | --------: |
| SustainedPerformance_Light_60Frames  | 100           | 126.2 us |  2.50 us |  5.00 us | 125.1 us |      64 B |
| SustainedPerformance_Medium_60Frames | 100           | 171.7 us |  3.31 us |  7.85 us | 167.6 us |      64 B |
| SustainedPerformance_Heavy_60Frames  | 100           | 353.4 us |  6.00 us | 12.39 us | 350.2 us |      64 B |
| SustainedPerformance_Light_60Frames  | 500           | 125.5 us |  2.49 us |  5.14 us | 124.2 us |     112 B |
| SustainedPerformance_Medium_60Frames | 500           | 170.6 us |  3.32 us |  7.50 us | 167.8 us |      64 B |
| SustainedPerformance_Heavy_60Frames  | 500           | 351.2 us |  6.85 us | 12.52 us | 347.6 us |     112 B |
| SustainedPerformance_Light_60Frames  | 1000          | 125.8 us |  2.47 us |  4.76 us | 124.7 us |      64 B |
| SustainedPerformance_Medium_60Frames | 1000          | 180.8 us |  3.58 us |  6.99 us | 181.6 us |     112 B |
| SustainedPerformance_Heavy_60Frames  | 1000          | 361.7 us |  7.05 us | 16.47 us | 357.3 us |      64 B |
| SustainedPerformance_Light_60Frames  | 2500          | 124.9 us |  2.19 us |  5.49 us | 122.2 us |     112 B |
| SustainedPerformance_Medium_60Frames | 2500          | 173.4 us |  3.27 us |  8.14 us | 171.0 us |      64 B |
| SustainedPerformance_Heavy_60Frames  | 2500          | 371.5 us | 11.37 us | 32.63 us | 362.4 us |     112 B |
| SustainedPerformance_Light_60Frames  | 5000          | 131.6 us |  5.26 us | 15.26 us | 124.2 us |      64 B |
| SustainedPerformance_Medium_60Frames | 5000          | 171.8 us |  3.10 us |  6.81 us | 169.3 us |     112 B |
| SustainedPerformance_Heavy_60Frames  | 5000          | 353.5 us |  6.81 us | 15.08 us | 348.8 us |      64 B |

## Key Technical Improvements

- Replaced linked lists with high-performance circular buffer using unmanaged memory
- Particle allocations and deallocations no longer require memory copying
- Lower memory allocations during runtime
- Particles are stored as value types with unsafe pointers for maximum performance.
- Sequential memory layout to improve CPU cache utilization.
- All particle operations work directly with unmanaged memory for higher performance
- Proper resource cleanup with finalizers and disposal tracking
- Comprehensive XML serialization and deserialization of particle effects.
- Dedicated classes for loading and saving particle configurations.
- Convenience methods for reading/writing helpers of common data types in XML
- Optimized rendering with better state management
- Now supports Front-to-back and Back-to-front particle rendering order from original Mercury particle engine.

## Memory Allocation Improvements

The new system shows consistently lower allocations:

- Current: 64-400B per frame.
- New: 64-122B per frame.

## Testing

- All current tests pass
- New tests added for serialization and deserialization to XML